backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
google-generativeai>=0.5.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
requests>=2.31.0
